phase2: fix TypeError in configuration save and batch manager
ceil() returns float in PHP but array_slice() expects ?int and array_fill()
expects int — with strict_types these now throw TypeError:

  admin/configuration.php:171 — array_slice(... ceil(...)) → (int) ceil(...)
  admin/batch_manager_global.php:50 — array_fill(0, $_POST['nb_photos_deleted'])
    string from POST passed as int; → (int) cast
  admin/batch_manager_unit.php:98 — same pattern

All three were silent before strict_types; the config save was blocking the
E2E spec-06 test. Baseline: 5017 → 5016.

+++ admin/configuration.php
@@ -168,7 +168,7 @@ if (isset($_POST['submit'])) {
                             $page['errors'][] = l10n('No order field selected');
                         } else {
                             // limit to the number of available parameters
-                            $order_by = $order_by_inside_category = array_slice($_POST['order_by'], 0, ceil(count($sort_fields) / 2));
+                            $order_by = $order_by_inside_category = array_slice($_POST['order_by'], 0, (int) ceil(count($sort_fields) / 2));
 
                             // there is no rank outside categories
                             if (($i = array_search('`rank` ASC', $order_by)) !== false) {
